Add assert equal tests.
