+++ plutus-core/plutus-core.cabal
@@ -499,12 +499,32 @@ executable traceToStacks
     import: lang
     main-is: Main.hs
     hs-source-dirs: executables/traceToStacks
+    other-modules:
+        Common
     build-depends:
         base >=4.9 && <5,
         cassava -any,
         integer-gmp -any,
         bytestring -any,
         optparse-applicative -any,
+        text -any,
+        vector -any,
+
+-- Tests for functions called by @traceToStacks@.
+test-suite traceToStacks-test
+    import: lang
+    type: exitcode-stdio-1.0
+    hs-source-dirs: executables/traceToStacks
+    default-language: Haskell2010
+    main-is: TestGetStacks.hs
+    other-modules:
+        Common
+    build-depends:
+        base -any,
+        bytestring -any,
+        cassava -any,
+        tasty -any,
+        tasty-hunit -any,
         text -any,
         vector -any,
